refactor: article, comment, like 훅 통합
